refactored model, eval, datasets, training
