+++ grazie/spell/main/data/datasets/bea/bea10.gt
@@ -1,0 +1,10 @@
+I WANT TO THANK YOU FOR PREPARING SUCH A GOOD PROGRAMME FOR US AND ESPECIALLY FOR TAKING US ON THE RIVER TRIP TO GREENWICH.
+IN MY OPINION FAMOUS PEOPLE ARE BEING OBLIGED TO PAY A PRICE FOR BEING FAMOUS THAT, IN SOME CASES, COSTS MORE THAN THEY DESERVE TO PAY.
+Also, I want to say that the plays and films were excellent, but there weren't enough of them for me.
+In our Academy we are not allowed to smoke.
+I was truly disappointed by it.
+Secondly, I had to wait forty - five minutes before the show finally began.
+I'd like you to send the money to this address: ul Taklowa 10
+It is a dream come true and was really unexpected for me!
+If not, what do you suggest?
+The festival was excellent in many ways, and in particular it being an international festival was a challenging, but brilliant idea.
