+++ grazie/spell/main/data/datasets/bea/bea10.noise
@@ -1,0 +1,10 @@
+I WANT TO THAK YOU FOR PREPARING SUCH A GOOD PROGRAMME FOR US AND ESPECIALLY FOR TAKING US ON THE RIVER TRIP TO GREENWICH.
+IN MY OPINION FAMOUS PEOPLE ARE BEING OBLIGED TO PAY A PRICE FOR BEING FAMOUS THAT, IN SOME CASS, COSTS MORE THAN THEY DESERVE TO PAY.
+Also, I want to say that the plays and films were exellent, but there weren't enough of them for me.
+In our Acadamy we are not allowed to smoke.
+I was trully dissapointed by it.
+Secondly, I had to wait fourty - five minutes before the show finally began.
+I'd like you to send the money to this adress: ul Taklowa 10
+It is a dream becames true and was really unexpected for me!
+If not, what do you sugest?
+The festival was excenent in many ways, and in particular it being an international festival was a challenging, but brilliant idea.
