+++ grazie/spell/main/data/experiments/char_based_transformer_big_9_epochs/report.json
@@ -1,0 +1,49 @@
+{
+    "Date": "25/03/2022 08:12",
+    "Model": "Char-Based Transformer, checkpoint: /home/ubuntu/omelnikov/grazie/spell/main/training/model_big_0_9.pt",
+    "Metrics": {
+        "Precision": 0.5,
+        "Recall": 0.8,
+        "F_0_5": 0.54,
+        "Word-level accuracy": 0.92,
+        "Broken tokenization cases": 0.4
+    },
+    "Mistakes examples": {
+        "Wrong correction of real mistake": [
+            {
+                "Text noise": "It is a dream becames true and was really unexpected for me!",
+                "Word noise": "becames",
+                "Word gt": "come",
+                "Word res": "becomes"
+            },
+            {
+                "Text noise": "Secondly, I had to wait fourty - five minutes before the show finally began.",
+                "Word noise": "fourty",
+                "Word gt": "forty",
+                "Word res": "fourth"
+            }
+        ],
+        "No mistake, but model corrected": [
+            {
+                "Text noise": "I'd like you to send the money to this adress: ul Taklowa 10",
+                "Word noise": "10",
+                "Word gt": "10",
+                "Word res": "100."
+            },
+            {
+                "Text noise": "I'd like you to send the money to this adress: ul Taklowa 10",
+                "Word noise": "ul",
+                "Word gt": "ul",
+                "Word res": "uk"
+            }
+        ],
+        "Not found mistake": [
+            {
+                "Text noise": "In our Acadamy we are not allowed to smoke.",
+                "Word noise": "Acadamy",
+                "Word gt": "Academy",
+                "Word res": "Acadamy"
+            }
+        ]
+    }
+}
